+++ app/rulesets/osu/performance/pp241007/evaluators/flashlight.go
@@ -3,7 +3,7 @@ package evaluators
 import (
 	"github.com/wieku/danser-go/app/beatmap/difficulty"
 	"github.com/wieku/danser-go/app/beatmap/objects"
-	"github.com/wieku/danser-go/app/rulesets/osu/performance/pp241009/preprocessing"
+	"github.com/wieku/danser-go/app/rulesets/osu/performance/pp241007/preprocessing"
 	"math"
 )
