+++ app/rulesets/osu/performance/pp241007/pp.go
@@ -1,8 +1,8 @@
-package pp241009
+package pp241007
 
 import (
 	"github.com/wieku/danser-go/app/beatmap/difficulty"
-	"github.com/wieku/danser-go/app/rulesets/osu/performance/pp241009/skills"
+	"github.com/wieku/danser-go/app/rulesets/osu/performance/pp241007/skills"
 	"github.com/wieku/danser-go/framework/math/mutils"
 	"math"
 )
