+++ app/rulesets/osu/performance/pp241007/skills/speed.go
@@ -2,8 +2,8 @@ package skills
 
 import (
 	"github.com/wieku/danser-go/app/beatmap/difficulty"
-	"github.com/wieku/danser-go/app/rulesets/osu/performance/pp241009/evaluators"
-	"github.com/wieku/danser-go/app/rulesets/osu/performance/pp241009/preprocessing"
+	"github.com/wieku/danser-go/app/rulesets/osu/performance/pp241007/evaluators"
+	"github.com/wieku/danser-go/app/rulesets/osu/performance/pp241007/preprocessing"
 	"math"
 )
